+++ src/track/track_utils.c
@@ -164,7 +164,7 @@ u16 tracker_load_cc_data(tracker_cc_data_t *cc_data) {
  * \return None
  */
 void tracker_set_carrier_phase_offset(const tracker_info_t *info,
-                                      double carrier_phase_offset) {
+                                      s64 carrier_phase_offset) {
   bool adjusted = false;
   tracker_t *tracker_channel = tracker_get(info->id);
 
@@ -181,7 +181,7 @@ void tracker_set_carrier_phase_offset(const tracker_info_t *info,
 
   if (adjusted) {
     log_debug_mesid(info->mesid,
-                    "Adjusting carrier phase offset to %lf",
+                    "Adjusting carrier phase offset to %" PRId64,
                     carrier_phase_offset);
   }
 }
