+++ src/utils/timing/timing.h
@@ -51,8 +51,8 @@ time_quality_t get_time_quality(void);
 void adjust_rcvtime_offset(const double dt);
 gps_time_t napcount2gpstime(const double tc);
 gps_time_t napcount2rcvtime(const double tc);
-double gpstime2napcount(const gps_time_t* t);
-double rcvtime2napcount(const gps_time_t* t);
+u64 gpstime2napcount(const gps_time_t* t);
+u64 rcvtime2napcount(const gps_time_t* t);
 u64 timing_getms(void);
 gps_time_t glo2gps_with_utc_params(me_gnss_signal_t mesid,
                                    const glo_time_t* glo_t);
